refactor: remove PipelineRunsProvider
